+++ src/stub/compare_arguments.hpp
@@ -5,44 +5,52 @@
 
 #pragma once
 
-#include "print_argument.hpp"
-
+#include <cstdint>
 #include <tuple>
 #include <type_traits>
+
+#include "compare_argument.hpp"
 
 namespace stub
 {
     /// Specialization chosen for empty tuples or when Index reaches the
     /// sizeof the tuple (i.e. the number of values in the tuple), see
     /// description below.
+    ///
+    /// @return When the two tuples are empty they compare equal so we return
+    ///         true
     template
     <
         class Index = std::integral_constant<uint32_t, 0U>,
         class... Args,
+        class... WithArgs,
         class LastIndex = std::integral_constant<uint32_t, sizeof...(Args)>,
         typename std::enable_if<
             std::is_same<Index,LastIndex>::value, uint8_t>::type = 0
     >
-    inline void print_arguments(std::ostream& out, const std::tuple<Args...>& t)
+    inline bool compare_arguments(const std::tuple<Args...>& actual,
+                                  const std::tuple<WithArgs...>& with)
     {
-        (void) out;
-        (void) t;
+        (void) actual;
+        (void) with;
+
+        return true;
     }
 
-    /// Prints the content of a tuple to the specified std::ostream.
+    /// Compare the content of two tuples.
     ///
-    /// The two functions print_arguments use SFINAE (Substitution Failure
+    /// The two functions compare_arguments use SFINAE (Substitution Failure
     /// Is Not An Error) to select which overload to call.
     ///
     /// The overloading works like this:
     ///
-    ///   1. If print_arguments is called with an empty tuple then the
+    ///   1. If compare_arguments is called with an empty tuple then the
     ///      empty overload will be chosen.
     ///
-    ///   2. If print_argument is called with a non-empty tuple the
-    ///      Index!=LastIndex is true and the overload writing to the
-    ///      std::ostream will be called. This will then recursively call
-    ///      print_arguments incrementing the Index. A some point
+    ///   2. If compare_arguments is called with a non-empty tuple the
+    ///      Index!=LastIndex is true and the overload calling
+    ///      compare_arugment(...) will be called. This will then recursively
+    ///      call compare_arguments incrementing the Index. A some point
     ///      Index==LastIndex and the empty overload gets chosen and we
     ///      are done.
     ///
@@ -54,15 +62,28 @@ namespace stub
     <
         class Index = std::integral_constant<uint32_t, 0U>,
         class... Args,
+        class... WithArgs,
         class LastIndex = std::integral_constant<uint32_t, sizeof...(Args)>,
         typename std::enable_if<
             !std::is_same<Index,LastIndex>::value, uint8_t>::type = 0
     >
-    inline void print_arguments(std::ostream& out, const std::tuple<Args...>& t)
+    inline bool compare_arguments(const std::tuple<Args...>& actual,
+                                  const std::tuple<WithArgs...>& with)
     {
-        print_argument(out, Index::value, std::get<Index::value>(t));
+        static_assert(sizeof...(Args) == sizeof...(WithArgs),
+            "The tuples must have same size");
 
-        print_arguments<
-            std::integral_constant<uint32_t, Index::value + 1>>(out, t);
+        bool result = compare_argument(std::get<Index::value>(actual),
+            std::get<Index::value>(with));
+
+        if (result == false)
+        {
+            return result;
+        }
+        else
+        {
+            using next = std::integral_constant<uint32_t, Index::value + 1>;
+            return compare_arguments<next>(actual, with);
+        }
     }
 }