+++ src/stub/function.hpp
@@ -10,28 +10,29 @@
 #include <ostream>
 
 #include "return_handler.hpp"
-#include "unqualified_type.hpp"
+//#include "unqualified_type.hpp"
 #include "print_arguments.hpp"
+#include "compare_call.hpp"
 
 namespace stub
 {
-    /// Default call
-    template<typename T> class call;
+    /// Default function
+    template<typename T> class function;
 
-    /// @brief The call object act like a "sink" for function calls
-    ///        i.e. we can define a call object to accept any type of
+    /// @brief The function object act like a "sink" for function calls
+    ///        i.e. we can define a function object to accept any type of
     ///        function call and it will simply store the arguments
     ///        for later inspection.
     ///
-    /// The typical use-case for the call object is when testing that
+    /// The typical use-case for the function object is when testing that
     /// some code invokes a specific set of functions with a specific
     /// set of arguments.
     ///
     /// Example:
     ///
-    ///    stub::call<void(uint32_t)> some_function;
+    ///    stub::function<void(uint32_t)> some_function;
     ///
-    /// The above call takes an uint32_t and returns nothing, lets
+    /// The above function takes an uint32_t and returns nothing, lets
     /// invoke it:
     ///
     ///     some_function(3);
@@ -45,9 +46,9 @@ namespace stub
     ///     bool called_with = some_function.expect_calls().with(4U);
     ///     assert(called_with == true);
     ///
-    /// We can also define a call which returns a value:
+    /// We can also define a function which returns a value:
     ///
-    ///     stub::call<bool(uint32_t)> another_function;
+    ///     stub::function<bool(uint32_t)> another_function;
     ///
     /// Here we have to specify what return value we expect:
     ///
@@ -63,57 +64,33 @@ namespace stub
     /// return_handler.hpp
     ///
     template<typename R, typename... Args>
-    class call<R (Args...)>
+    class function<R (Args...)>
     {
     public:
 
-        /// Get the unqualified type, i.e. if the function takes a
-        /// reference to some type we make sure we store the actual type.
-        ///
-        /// Example:
-        ///
-        ///     stub::call<void(const int&)> function;
-        ///     function(3);
-        ///
-        /// In this case we will store an int of value 3 instead of a
-        /// const& to the temporary value which will no longer exist
-        /// when we want to compare.
-        ///
-        /// A tuple is used to store the arguments passed
-        using arguments = std::tuple<typename unqualified_type<Args>::type...>;
-
-        /// The default binary predicate type use when comparing
-        /// function calls
-        using default_predicate =
-            std::function<bool(const arguments&,const arguments&)>;
-
-    public:
-
-        /// Represent a expectation of how the call object has been
+        /// Represent a expectation of how the function object has been
         /// invoked. Using the API it is possible to setup how we
-        /// expect the call object looks like. The expectation
+        /// expect the function object looks like. The expectation
         /// converts to bool allowing the user to detect whether the
         /// expectation was correct.
-        template<class BinaryPredicate>
         struct expectation
         {
-            /// @param the_call The call we configuring an expectation for
+            /// @param the_function The function we configuring an expectation for
             ///
             /// @param predicate The function object used to compare the
             ///        call arguments
-            expectation(const call& the_call, const BinaryPredicate& predicate)
-                : m_call(the_call),
-                  m_predicate(predicate)
+            expectation(const function& the_function)
+                : m_function(the_function)
             { }
 
             /// Calling with(...) will add a set of arguments we
             /// expect to see. with(...) can be called multiple times
             /// in a row if we expect multiple function calls to the
-            /// call object.
+            /// function object.
             ///
             /// As an example:
             ///
-            ///     stub::call<void(uint32_t,uint32_t)> function;
+            ///     stub::function<void(uint32_t,uint32_t)> function;
             ///     function(3,1);
             ///     function(4,2);
             ///
@@ -125,9 +102,10 @@ namespace stub
             ///
             /// @return The expectation itself, which allows chaining
             ///         function calls
-            expectation& with(Args... args)
+            template<class... WithArgs>
+            expectation& with(WithArgs&&... args)
             {
-                m_calls.emplace_back(args...);
+                m_calls.emplace_back(std::forward<WithArgs>(args)...);
                 return *this;
             }
 
@@ -146,19 +124,25 @@ namespace stub
                 // An expectation can't be evaluated if it hasn't been setup.
                 assert(!m_calls.empty());
 
-                if (m_call.m_calls.size() != m_calls.size())
+                if (m_function.m_calls.size() != m_calls.size())
                     return false;
 
-                return std::equal(std::begin(m_call.m_calls),
-                                  std::end(m_call.m_calls),
-                                  std::begin(m_calls),
-                                  m_predicate);
+                for (uint32_t i = 0; i < m_calls.size(); ++i)
+                {
+                    auto& actual = m_function.m_calls[i];
+                    if (m_calls[i].compare(actual) == false)
+                    {
+                        return false;
+                    }
+                }
+
+                return true;
             }
 
             /// Use the to_bool member function when casting thís expectation
             /// to a boolean value.
             ///
-            /// @return True if the expectation matches the call,
+            /// @return True if the expectation matches the function,
             ///         otherwise false
             explicit operator bool() const
             {
@@ -167,15 +151,11 @@ namespace stub
 
         private:
 
-            /// The call we will check the expectation against
-            const call& m_call;
-
-            /// The comparison function used when comparing whether
-            /// the function call arguments match
-            BinaryPredicate m_predicate;
+            /// The function we will check the expectation against
+            const function& m_function;
 
             /// The expected calls
-            std::vector<arguments> m_calls;
+            std::vector<compare_call<Args...>> m_calls;
         };
 
     public:
@@ -186,89 +166,62 @@ namespace stub
         /// @return The return value generated by the return_handler
         R operator()(Args... args) const
         {
-            m_calls.emplace_back(args...);
+            m_calls.emplace_back(std::forward<Args>(args)...);
             return m_return_handler();
         }
 
         /// @copydoc return_handler::set_return(const T&)
-        template<class T>
-        return_handler<R>& set_return(const T& return_value)
+        template<class... Returns>
+        return_handler<R>& set_return(Returns&&... return_value)
         {
-            return m_return_handler.set_return(return_value);
-        }
-
-        /// @copydoc return_handler::set_return(
-        ///              const std::initializer_list<T>&)
-        template<class T>
-        return_handler<R> set_return(const std::initializer_list<T> &returns)
-        {
-            return m_return_handler.set_return(returns);
+            return m_return_handler.set_return(
+                std::forward<Returns>(return_value)...);
         }
 
         /// @return The number of times the call operator has been invoked
         uint32_t calls() const
         {
-            return (uint32_t)m_calls.size();
-        }
-
-        /// @return True if no calls have been made otherwise false
-        bool no_calls() const
-        {
-            return m_calls.size() == 0;
+            return (uint32_t) m_calls.size();
         }
 
         /// @return The arguments passed to the n'th call
-        const arguments& call_arguments(uint32_t index) const
+        const arguments<Args...>& call_arguments(uint32_t index) const
         {
             assert(index < m_calls.size());
             return m_calls[index];
         }
 
-        /// Used when we want to check whether the call object is in a
+        /// Used when we want to check whether the function object is in a
         /// certain state. See examples usage in the expectation
         /// struct member functions.
         ///
-        /// Two version of this function exists one which takes a
-        /// binary predicate comparison function an one which does
-        /// not. The version taking a predicate allows the user to
-        /// provide a custom comparison function whereas the other
-        /// version will rely on operator==(...).
-        ///
-        /// When you pass your own predicate function it will get the
-        /// actual call arguments as the first parameter and the
-        /// expected as the second.
-        ///
         /// @return An expectation object
-        expectation<default_predicate> expect_calls() const
+        expectation expect_calls() const
         {
-            default_predicate predicate =
-                [](const arguments& a, const arguments& b) -> bool
-                   { return a == b; };
-
-            return expectation<default_predicate>(*this, predicate);
+            return expectation(*this);
         }
 
-        /// @copydoc expect_calls()
-        template<class BinaryPredicate>
-        expectation<BinaryPredicate> expect_calls(
-            const BinaryPredicate& predicate) const
+        /// Removes all calls from the function object and reset the return
+        /// handler.
+        void clear()
         {
-            return expectation<BinaryPredicate>(*this, predicate);
+            m_return_handler = return_handler<R>();
+            m_calls.clear();
         }
 
-        /// Prints the status of the call object to the std::ostream.
+        /// Prints the status of the function object to the std::ostream.
         ///
         /// Example (using the output operator):
         ///
-        ///    stub::call<void(uint32_t)> my_func;
+        ///    stub::function<void(uint32_t)> my_func;
         ///
         ///    my_func(4U);
         ///    my_func(5U);
         ///
-        ///    // Print the current status of the call object,
+        ///    // Print the current status of the function object,
         ///    std::cout << my_func << std::endl;
         ///
-        /// @param out The ostream where the stub::call status should be
+        /// @param out The ostream where the stub::function status should be
         void print(std::ostream& out) const
         {
             out << "Number of calls: " << m_calls.size() << std::endl;
@@ -289,23 +242,23 @@ namespace stub
         return_handler<R> m_return_handler;
 
         /// Stores the arguments every time the operator() is invoked
-        mutable std::vector<arguments> m_calls;
+        mutable std::vector<arguments<Args...>> m_calls;
     };
 
-    /// Output operator for printing call objects, see more info in
-    /// stub::call::print(std::ostream&).
+    /// Output operator for printing function objects, see more info in
+    /// stub::function::print(std::ostream&).
     ///
     ///
-    /// @param out The output stream where the state of the call object
+    /// @param out The output stream where the state of the function object
     ///        will be printed.
     ///
-    /// @param call The call object we want to print
+    /// @param function The function object we want to print
     ///
     /// @return The ostream operator.
     template<class T>
-    inline std::ostream& operator<<(std::ostream& out, const call<T>& call)
+    inline std::ostream& operator<<(std::ostream& out, const function<T>& function)
     {
-        call.print(out);
+        function.print(out);
         return out;
     }