+++ test/src/test_ignore.cpp
@@ -1,0 +1,10 @@
+// Copyright (c) 2016 Steinwurf ApS
+// All Rights Reserved
+//
+// Distributed under the "BSD License". See the accompanying LICENSE.rst file.
+
+#include <stub/ignore.hpp>
+
+#include <gtest/gtest.h>
+
+// stub::ignore is just a tag object - so no unit test to run.