+++ test/src/test_return_handler.cpp
@@ -7,14 +7,14 @@
 
 #include <gtest/gtest.h>
 
-TEST(return_handler, void)
+TEST(test_return_handler, void)
 {
     // Just checking that it can be instantiated
     stub::return_handler<void> r;
     r();
 }
 
-TEST(return_handler, api)
+TEST(test_return_handler, api)
 {
     // Return a single value multiple times
     {
@@ -29,7 +29,7 @@ TEST(return_handler, api)
     // Use the return value of set_return
     {
         stub::return_handler<uint32_t> r;
-        r.set_return({5U, 3U});
+        r.set_return(5U, 3U);
 
         EXPECT_EQ(5U, r());
         EXPECT_EQ(3U, r());
@@ -45,7 +45,7 @@ TEST(return_handler, api)
 
         EXPECT_EQ(5U, r());
 
-        r.set_return({3U,4U}).no_repeat();
+        r.set_return(3U,4U).no_repeat();
 
         EXPECT_EQ(3U, r());
         EXPECT_EQ(4U, r());
@@ -54,12 +54,21 @@ TEST(return_handler, api)
     // Death tests
     {
         stub::return_handler<uint32_t> r;
-        r.set_return({3U,4U}).no_repeat();
+        r.set_return(3U,4U).no_repeat();
 
         EXPECT_EQ(3U, r());
         EXPECT_EQ(4U, r());
 
         // death
         // EXPECT_EQ(r(), 4U);
+    }
+
+    // String
+    {
+        stub::return_handler<const char*> r;
+        r.set_return("3U","4U").no_repeat();
+
+        EXPECT_EQ("3U", r());
+        EXPECT_EQ("4U", r());
     }
 }