+++ test/src/test_unqualified_type.cpp
@@ -7,6 +7,9 @@
 
 #include <gtest/gtest.h>
 
+// @todo remove
+#include <tuple>
+
 // Small alias to make testing easier
 
 namespace
@@ -16,7 +19,7 @@ namespace
 }
 
 /// Test that the unqualified type helper works as expected
-TEST(unqualified_type, test)
+TEST(test_unqualified_type, test)
 {
     {
         auto ok = test<const uint32_t&, uint32_t>::value;